requirements.txt
botocore~=1.21.56
boto3~=1.18.56
pandas~=1.3.3
SQLAlchemy~=1.4.25
python-dotenv~=0.19.0